Refactor ShareSnippet management by restructuring resources
- Moved ShareSnippetResource to a new namespace: Firefly\FilamentBlog\Resources\ShareSnippets.
- Introduced new schemas and pages for managing share snippets, including CreateShareSnippet, EditShareSnippet, and ListShareSnippets.
- Removed the deprecated ShareSnippetResource class and its associated methods to enhance clarity and maintainability.
- Updated Blog.php and other related files to reflect the new ShareSnippetResource path.
