+++ src/Blog.php
@@ -8,7 +8,7 @@ use Firefly\FilamentBlog\Resources\Tags\TagResource;
 use Firefly\FilamentBlog\Resources\SeoDetails\SeoDetailResource;
 use Firefly\FilamentBlog\Resources\Newsletters\NewsletterResource;
 use Firefly\FilamentBlog\Resources\Comments\CommentResource;
-use Firefly\FilamentBlog\Resources\ShareSnippetResource;
+use Firefly\FilamentBlog\Resources\ShareSnippets\ShareSnippetResource;
 use Firefly\FilamentBlog\Resources\SettingResource;
 use Filament\Contracts\Plugin;
 use Filament\Panel;
