+++ src/Resources/ShareSnippets/Tables/ShareSnippetsTable.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace Firefly\FilamentBlog\Resources\ShareSnippets\Tables;
+
+use Filament\Actions\BulkActionGroup;
+use Filament\Actions\DeleteBulkAction;
+use Filament\Actions\EditAction;
+use Filament\Tables\Columns\TextColumn;
+use Filament\Tables\Columns\ToggleColumn;
+use Filament\Tables\Table;
+
+class ShareSnippetsTable
+{
+    public static function configure(Table $table): Table
+    {
+        return $table
+            ->columns([
+                TextColumn::make('script_code')
+                    ->limit(50)
+                    ->searchable(),
+                TextColumn::make('html_code')
+                    ->limit(50)
+                    ->searchable(),
+                ToggleColumn::make('active'),
+            ])
+            ->recordActions([
+                EditAction::make(),
+            ])
+            ->toolbarActions([
+                BulkActionGroup::make([
+                    DeleteBulkAction::make(),
+                ]),
+            ]);
+    }
+}
